Refactor ws-server component.
The previous websocket-server component was used by an
example, and contained some hard-coded information
to be used in tests. It is replaced with a generic
ws-server component.

The generic component accepts a socket that is used
as a sink. The socket has to have a `send` method
and a `message` and `close` event.