Allow empty pipelines.
It's not necessary to have at least 1 component
in the pipeline, as all methods are accessed via
existing components only, so there does not need
to be a first of last component.

This is needed to have pipelines that can be initialized
later, as the components might not be known during
initialization.