Remove custom handler on MSE component.
We had added custom handler on the MSE component
in order to be able to call functions for specific
message types to be able to synchronize different
messages.

However, a better solution is to couple such handlers
to specific components that already deal with those
message types.

Synchronization will happen instead by sending an
event from the MP4 muxer component when presentation
time is known.