Send notification when presentation time is known.
The MP4 muxer will call the `onSync` method whenever
the presentation time has been determined, i.e. the
timestampe corresponding to the beginning of the
presentation.

The video pipeline that contains the MP4 muxer also
gets the same `onSync` method, that will be called
whenever the presentation time is known.