[breaking] Improved examples.
Instead of videoEl and canvasEl, the generic `mediaElement`
property is used to configure the element for use in either
the mse or canvas components. This breaks compatibility with
previous use of the HtmlVideoPipeline and HtmlCanvasPipeline.

Examples changes:

 - examples have been restructured so that it's easier
   to serve them from one place
 - better RTSP server control for test purposes (separate
   script and additional text overlays to indicate encoding)
 - examples for MJPEG were added

Control the RTSP/WebSocket servers with a dedicated script.

Instead of including a ./run.sh script for every example,
the script to manage the RTSP and WebSocket servers is
moved to a dedicated directory, and can be launched from
anywhere in the tree structure, or using yarn rtsp-ws-server.

The Travis configuration is updated to point to the new
tcp-ws-proxy and example directories, and the cypress
test points to `test/h264.html` explicitely.