fix: only clear buffer before last I-frame
When removing buffered video data, there needs to be
at least an I-frame left in the remaining buffer, otherwise
the video playback freezes (both chrome and firefox).

With this change, we mark I-frames and make sure that there
is at least one I-frame in the remaining buffer when clearing
old buffered data.