chore: refactor H.264 depay as a class
Instead of keeping the buffer outside of the parser,
use a class instead which keeps track of the buffer
internally.

The `.parse` method on the class either returns a
new H.264 message, or null (in case there is no message
yet). In all cases, it keeps an internal buffer with
data that's part of the next message.

This leverages the parser's logic to determine if
the message is an I-frame in order to determine the
start of the stream (instead of coding it explicitly
in the transform stream).