fix: remove unnecessary context value changes
Especially Liner, but also Foundation, sent out a new context value
each time the component was rendered.

This adds a useMemo around areaBBox so that the value and functions in
the Liner context don't change every render. It also adds useMemo for
the actual context value for both Linder and Foundation.