feat!: expose browser by default
Trying to have default entry points for both
node and browser seems to be a bit of a mess
to have it work properly for all kinds of
environments.

Instead, only browser entry points are exposed,
with the `browser` field pointing to the bundle
(so you don't need to be able to bundle the
browserify parts yourself when using this).

If using with Node.js, you'll have to import
from `dist/cjs/index.node.js` directly.