feat!: default support for modern browsers
Using async/await causes issues with supporting
older browsers, in that we need to include regenerator
runtime. In order to keep things as simple as possible
and the resulting bundle as small as possible, the
browserslist is set explicitely to modern browsers.

Instead, a second "legacy" bundle is now available
for those using e.g. IE11 (or another older browser).

BREAKING CHANGES:
  - if you support sufficiently modern browsers
    you shouldn't need to change anything
  - if you support older browsers, use the new bundle
    `media-stream-library.legacy.min.js` instead