feat: pass explicit host to TCP component
Instead of parsing the RTSP URI in the TCP component
to determine the actual hostname and port, allow to
pass a host to the component that will be used as an
override (ignoring the RTSP URI host).

This is useful in cases where the client doesn't know
the actual RTSP server host.