feat!: optional RTSP host for TcpWsProxyPipeline
The TcpSource now supports an optional RTSP host,
so the configuration of the TcpWsProxyPipeline is
adjusted to allow passing such an option.

The tcp-ws-proxy cli now has an --rtspHost option
to pass an RTSP host to be used instead of the URI
from the RTSP OPTIONS command. This allows us to
hardwire the RTSP server location in the WebSocket
proxy itself, so the client doesn't need to know
about it.

BREAKING CHANGES:
- previous config to TcpWsProxyPipeline is now
  passed as TcpWsProxyPipeline({wsConfig}) instead