chore(deps): upgrade all dependencies
Legacy build for IE11 was removed as it no longer worked anyway with the
latest version of `buffer` (see #632). Instead, a separate `ie11` branch
will be created that contains the legacy bundle.

With that, all webpack/babel related packages are now removed.

The spurious `md5.js` package was also removed, closes #660.