chore!: migrate from webpack to esbuild/vite and upgrade dependencies.
Bundling is done with esbuild instead of webpack, and development is
done with vite instead of webpack.

Pnp is removed as this does not work with esbuild (but deps are still
committed for no-network-zero-install).

BREAKING: when importing the video player as a module, the import
path will now provide the cjs/esm bundles without external dependencies
bundled. This means your bundler should take care of this. It might be that
you need to add a definition to replace global with window to make it work.