chore: add events explicitely (#672)
The stream-browserify package should "just" work in the browser
but it wraps the readable-stream package which is targeting node.
A bunch of work-arounds are needed to get this to work nicely for
client code that uses this library. In this case, we need to specify
that the events package is needed, as readable-stream does not list
that as a dependency.