chore: use yarn and replace dprint/eslint with biome
Fairly large maintenance-related changes, but no code
changes, aimed at simplifying maintenance.

Unfortunately, while the change to npm was an attempt
at using more default tooling, the additional burden
when updating packages, and the need to rely on a
third-party tool (bringing in lots of new dependencies)
turned out to be not great.
This re-introduces yarn, but it's kept at v3 (instead
of latest stable v4), because of dependabot support.

Another high-maintenance area is linting, with eslint
needing extensive configuration. This is now replaced
with biome. And because the latter also has a formatter,
dprint is also removed.