feat(streams): capture MP4 stream
Adds a `capture` method to the pipelines that produce MP4 data.
It stores all data in a buffer at start of movie, and passes it to
the provided callback when capture ends. The latter can be triggered
by calling the returned trigger function, or automatically when the
buffer is full.

Co-authored-by: Rikard Tegnander <rikardk@axis.com>
Co-authored-by: Victor Ingvarsson <victor.ingvarsson@axis.com>