feat(streams): allow changing the RTSP URI
Adds a URI option to the describe method in the RTSP session.
If provided, it's used for that session instead of the URI that
was provided to the constructor. The latter now only serves as
default URI when none is provided to describe.
On teardown, the session control URL is restored to the default.

Fixes #1077