chore(player): remove styled-components & update style
The major version of styled-components has been lagging behind and to
avoid further issues with React compatibility, it's no longer required
to have this as a peer dependency.

Examples are fine using inline style, and in other cases the React
`style` tag is used instead.

Additionally, the stats section is restyled to be much more simple.